Add acmeclient_test system test.
Currently does not test tls-sni-01 if http-01 is allowed, since apachesolver prefers http-01.

This requires a heavily reconfigured boulder test environment in which we use a user-defined bridge network instead of binding to the Docker host. This allows better
isolation and means less hazzle for host lookups. Docker doesn't set hostnames with TLDs, so that has to be patched away.